Prebacen end point iz UserFilescontroller-a u DocumentsController za dobijanje svih documenta jedog user-a
